requirements.txt
absl-py==0.8.1
appdirs==1.4.3
astor==0.8.0
astroid==2.3.3
attrs==19.3.0
black==19.10b0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
importlib-metadata==0.23
isort==4.3.21
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==19.2
pathspec==0.6.0
pluggy==0.13.1
protobuf==3.10.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pylint==2.4.4
pyparsing==2.4.5
pytest==5.3.0
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.0
six==1.13.0
tensorboard==2.0.1
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
